lock_service/requirements.txt
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
psycopg2==2.7.3
requests